refactor(sybra): extract skillsync package
Move syncSkills, syncFile, syncSkillsDir, syncFSSkillsDir,
copyDirTree, copySkillDir, hasYAMLFrontmatter (~330 LOC)
out of internal/sybra/app.go into a new internal/skillsync
package. App keeps a thin syncSkillsBundle adapter that
plumbs RepoDir, SkillsFS, dst paths into skillsync.Run.

Tests follow the code: 11 sync tests move from app_test.go
to skillsync/syncer_test.go and instantiate Syncer directly
(no App fixture needed).

Step 1/3 of issue 604.
