+++ internal/sybra/app_test.go
@@ -367,329 +367,6 @@ func TestGetAgentOutputNotFound(t *testing.T) {
 	}
 }
 
-func TestSyncFile(t *testing.T) {
-	a := setupApp(t)
-	a.repoDir = t.TempDir()
-
-	srcDir := filepath.Join(a.repoDir, "sub")
-	if err := os.MkdirAll(srcDir, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	srcFile := filepath.Join(srcDir, "test.md")
-	if err := os.WriteFile(srcFile, []byte("# hello"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	dstFile := filepath.Join(t.TempDir(), "out", "test.md")
-	a.syncFile(srcFile, dstFile)
-
-	data, err := os.ReadFile(dstFile)
-	if err != nil {
-		t.Fatalf("dst not written: %v", err)
-	}
-	if string(data) != "# hello" {
-		t.Errorf("content = %q, want %q", string(data), "# hello")
-	}
-}
-
-func TestSyncFileMissingSrc(t *testing.T) {
-	a := setupApp(t)
-	dstFile := filepath.Join(t.TempDir(), "should-not-exist.md")
-	a.syncFile("/nonexistent/file.md", dstFile)
-
-	if _, err := os.Stat(dstFile); !os.IsNotExist(err) {
-		t.Error("dst should not be created when src missing")
-	}
-}
-
-func TestSyncSkillsDir(t *testing.T) {
-	a := setupApp(t)
-
-	srcDir := filepath.Join(t.TempDir(), "skills")
-	if err := os.MkdirAll(srcDir, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	frontmatter := func(name string) []byte {
-		return []byte("---\nname: " + name + "\ndescription: test\n---\n\n# " + name)
-	}
-	for _, name := range []string{"a", "b"} {
-		if err := os.WriteFile(filepath.Join(srcDir, name+".md"), frontmatter(name), 0o644); err != nil {
-			t.Fatal(err)
-		}
-	}
-	// Non-.md, and a malformed .md without frontmatter must be skipped.
-	if err := os.WriteFile(filepath.Join(srcDir, "c.txt"), []byte("content-c"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(srcDir, "no-frontmatter.md"), []byte("# nope"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	dstDir := filepath.Join(t.TempDir(), "dst-skills")
-	a.syncSkillsDir(srcDir, dstDir)
-
-	// Each valid flat skill must land as <name>/SKILL.md, which is the layout
-	// Claude Code and Codex both require — flat .md files in the destination
-	// are never discovered by the skill loader.
-	for _, name := range []string{"a", "b"} {
-		skillMD := filepath.Join(dstDir, name, "SKILL.md")
-		if _, err := os.Stat(skillMD); err != nil {
-			t.Errorf("%s/SKILL.md missing at dst: %v", name, err)
-		}
-	}
-	if _, err := os.Stat(filepath.Join(dstDir, "no-frontmatter")); !os.IsNotExist(err) {
-		t.Errorf("malformed skill should not be written: stat err=%v", err)
-	}
-}
-
-func TestSyncSkillsDirRemovesOrphans(t *testing.T) {
-	a := setupApp(t)
-
-	srcDir := filepath.Join(t.TempDir(), "skills")
-	if err := os.MkdirAll(srcDir, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	keep := []byte("---\nname: keep\ndescription: test\n---\n\n# keep")
-	if err := os.WriteFile(filepath.Join(srcDir, "keep.md"), keep, 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	dstDir := filepath.Join(t.TempDir(), "dst-skills")
-	orphan := filepath.Join(dstDir, "orphan")
-	if err := os.MkdirAll(orphan, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(orphan, "SKILL.md"), []byte("---\nname: orphan\n---\n"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	a.syncSkillsDir(srcDir, dstDir)
-
-	if _, err := os.Stat(orphan); !os.IsNotExist(err) {
-		t.Errorf("orphan skill dir should be removed: stat err=%v", err)
-	}
-	if _, err := os.Stat(filepath.Join(dstDir, "keep", "SKILL.md")); err != nil {
-		t.Errorf("keep/SKILL.md missing: %v", err)
-	}
-}
-
-func TestSyncSkillsDirCopiesDirectoryStyleSkill(t *testing.T) {
-	a := setupApp(t)
-
-	srcDir := filepath.Join(t.TempDir(), "skills")
-	skillDir := filepath.Join(srcDir, "plan-critic")
-	refDir := filepath.Join(skillDir, "references")
-	if err := os.MkdirAll(refDir, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	skillMD := []byte("---\nname: plan-critic\ndescription: test\n---\n\n# plan-critic")
-	if err := os.WriteFile(filepath.Join(skillDir, "SKILL.md"), skillMD, 0o644); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(refDir, "checklist.md"), []byte("# checklist"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	badDir := filepath.Join(srcDir, "bad-skill")
-	if err := os.MkdirAll(badDir, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(badDir, "SKILL.md"), []byte("# no frontmatter"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	dstDir := filepath.Join(t.TempDir(), "dst-skills")
-	a.syncSkillsDir(srcDir, dstDir)
-
-	if _, err := os.Stat(filepath.Join(dstDir, "plan-critic", "SKILL.md")); err != nil {
-		t.Errorf("SKILL.md missing at dst: %v", err)
-	}
-	if _, err := os.Stat(filepath.Join(dstDir, "plan-critic", "references", "checklist.md")); err != nil {
-		t.Errorf("nested reference missing at dst: %v", err)
-	}
-	if _, err := os.Stat(filepath.Join(dstDir, "bad-skill")); !os.IsNotExist(err) {
-		t.Errorf("bad skill dir should not be copied: stat err=%v", err)
-	}
-}
-
-func TestSyncSkillsDirDirectoryStyleAndFrontmatter(t *testing.T) {
-	a := setupApp(t)
-
-	srcDir := filepath.Join(t.TempDir(), "skills")
-	if err := os.MkdirAll(srcDir, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	// Valid flat skill.
-	flat := []byte("---\nname: good-flat\ndescription: t\n---\n\n# flat")
-	if err := os.WriteFile(filepath.Join(srcDir, "good-flat.md"), flat, 0o644); err != nil {
-		t.Fatal(err)
-	}
-	// Malformed flat skill — must NOT overwrite a valid dst skill of same name.
-	if err := os.WriteFile(filepath.Join(srcDir, "bad-flat.md"), []byte("# no-fm"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	// Valid directory-style skill.
-	dirSkill := filepath.Join(srcDir, "good-dir")
-	if err := os.MkdirAll(dirSkill, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	dirMD := []byte("---\nname: good-dir\ndescription: t\n---\n\n# dir")
-	if err := os.WriteFile(filepath.Join(dirSkill, "SKILL.md"), dirMD, 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	dstDir := filepath.Join(t.TempDir(), "codex-skills")
-	// Pre-populate a valid bad-flat entry in the dst to verify it survives.
-	preExisting := filepath.Join(dstDir, "bad-flat")
-	if err := os.MkdirAll(preExisting, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	preExistingContent := []byte("---\nname: bad-flat\ndescription: valid\n---\n\n# valid")
-	if err := os.WriteFile(filepath.Join(preExisting, "SKILL.md"), preExistingContent, 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	a.syncSkillsDir(srcDir, dstDir)
-
-	if _, err := os.Stat(filepath.Join(dstDir, "good-flat", "SKILL.md")); err != nil {
-		t.Errorf("good-flat not written: %v", err)
-	}
-	if _, err := os.Stat(filepath.Join(dstDir, "good-dir", "SKILL.md")); err != nil {
-		t.Errorf("good-dir not written: %v", err)
-	}
-	// bad-flat must be treated as orphan (invalid src) and removed, OR preserved
-	// unchanged. The key invariant: dst must not contain the malformed payload.
-	got, err := os.ReadFile(filepath.Join(dstDir, "bad-flat", "SKILL.md"))
-	if err == nil && string(got) == "# no-fm" {
-		t.Errorf("malformed skill overwrote valid dst: %q", got)
-	}
-}
-
-func TestSyncSkillsPrefersEmbeddedWhenNoGoMod(t *testing.T) {
-	a := setupApp(t)
-	// Use in-memory FS stand-in: point skillsFS at a real testdata dir.
-	embeddedSrc := filepath.Join(t.TempDir(), "embedded")
-	dataDir := filepath.Join(embeddedSrc, "data")
-	if err := os.MkdirAll(dataDir, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	embedded := []byte("---\nname: embedded-skill\ndescription: from embed\n---\n\n# embed")
-	if err := os.WriteFile(filepath.Join(dataDir, "embedded-skill.md"), embedded, 0o644); err != nil {
-		t.Fatal(err)
-	}
-	a.skillsFS = os.DirFS(embeddedSrc)
-
-	// repoDir with a skills dir but NO go.mod — simulates container where
-	// cwd=/home/sybra. The embedded bundle must take precedence even though
-	// the skills dir already exists with a rogue file.
-	repoDir := t.TempDir()
-	a.repoDir = repoDir
-	skillsSrc := filepath.Join(repoDir, ".claude", "skills")
-	if err := os.MkdirAll(skillsSrc, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(skillsSrc, "rogue.md"), []byte("# no-fm"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	a.skillsDir = filepath.Join(t.TempDir(), "app-skills")
-
-	a.syncSkills()
-
-	// Embedded content lands at <name>/SKILL.md — the layout Claude Code
-	// and Codex require. Flat .md files at the top are not discoverable.
-	if _, err := os.Stat(filepath.Join(a.skillsDir, "embedded-skill", "SKILL.md")); err != nil {
-		t.Errorf("embedded skill missing: %v", err)
-	}
-	if _, err := os.Stat(filepath.Join(a.skillsDir, "rogue")); !os.IsNotExist(err) {
-		t.Errorf("rogue skill leaked into app skills dir")
-	}
-}
-
-func TestSyncSkillsUsesRepoSourceWhenGoModPresent(t *testing.T) {
-	a := setupApp(t)
-
-	repoDir := t.TempDir()
-	a.repoDir = repoDir
-	// Marker that this is the actual sybra source repo.
-	if err := os.WriteFile(filepath.Join(repoDir, "go.mod"), []byte("module x\n"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	skillsSrc := filepath.Join(repoDir, ".claude", "skills")
-	if err := os.MkdirAll(skillsSrc, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	repoSkill := []byte("---\nname: repo-skill\ndescription: t\n---\n")
-	if err := os.WriteFile(filepath.Join(skillsSrc, "repo-skill.md"), repoSkill, 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	// Even with an embedded FS available, go.mod presence selects repo source.
-	embeddedSrc := filepath.Join(t.TempDir(), "embedded")
-	dataDir := filepath.Join(embeddedSrc, "data")
-	if err := os.MkdirAll(dataDir, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(dataDir, "embed-only.md"), []byte("---\nname: embed-only\n---\n"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	a.skillsFS = os.DirFS(embeddedSrc)
-	a.skillsDir = filepath.Join(t.TempDir(), "app-skills")
-
-	a.syncSkills()
-
-	if _, err := os.Stat(filepath.Join(a.skillsDir, "repo-skill", "SKILL.md")); err != nil {
-		t.Errorf("repo skill missing: %v", err)
-	}
-	if _, err := os.Stat(filepath.Join(a.skillsDir, "embed-only")); !os.IsNotExist(err) {
-		t.Errorf("embedded skill should not be written in repo-source mode")
-	}
-}
-
-func TestSyncSkillsDirMissingSrc(t *testing.T) {
-	a := setupApp(t)
-	dstDir := filepath.Join(t.TempDir(), "should-not-exist")
-	a.syncSkillsDir("/nonexistent/dir", dstDir)
-
-	if _, err := os.Stat(dstDir); !os.IsNotExist(err) {
-		t.Error("dst dir should not be created when src missing")
-	}
-}
-
-func TestSyncSkillsNoRepoDir(t *testing.T) {
-	a := setupApp(t)
-	a.repoDir = ""
-	// Should not panic; falls back to cwd
-	a.syncSkills()
-}
-
-func TestSyncSkillsWithRepoDir(t *testing.T) {
-	a := setupApp(t)
-
-	repoDir := t.TempDir()
-	a.repoDir = repoDir
-
-	// Create source skills dir
-	skillsSrc := filepath.Join(repoDir, ".claude", "skills")
-	if err := os.MkdirAll(skillsSrc, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	skillContent := []byte("---\nname: skill\ndescription: test\n---\n\n# skill")
-	if err := os.WriteFile(filepath.Join(skillsSrc, "skill.md"), skillContent, 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	// Create orchestrator CLAUDE.md
-	orchDir := filepath.Join(repoDir, "orchestrator")
-	if err := os.MkdirAll(orchDir, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(orchDir, "CLAUDE.md"), []byte("# orchestrator"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	// Should not panic
-	a.syncSkills()
-}
-
 func TestShutdown(t *testing.T) {
 	a := setupApp(t)
 	a.Shutdown(t.Context())
